= separate database management from data access added Dao, RelTypes enum, db.Constants Facts, Entities, Relationships now use dao instead of db
